backend/pyproject.toml
[project]
name = "odontoflow"
version = "0.1.0"
description = "Sistema integrado de gestao odontologica"
requires-python = ">=3.9"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Validation & serialization
    "pydantic>=2.10",
    "pydantic[email]>=2.10",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # Utils
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
    # Storage
    "boto3>=1.35",
    # Redis
    "redis>=5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.8",
]
communication = [
    "httpx>=0.28",  # WhatsApp Business API
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["odontoflow"]
